+++ docs/Whats-New.md
@@ -47,6 +47,7 @@ New:
 - Unlimited Global / Local Variables (by secsome)
 - Adds a "Load Game" button to the retry dialog on mission failure (by secsome)
 - Default disguise for individual InfantryTypes (by secsome)
+- Fixed the bug that units' lighting get corrupted after loading a save with a different lighting being set (by secsome)
 - Allow `NotHuman=yes` infantry to use random `Death` anim sequence (by Otamaa)
 - Ability for warheads to trigger specific `NotHuman=yes` infantry `Death` anim sequence (by Otamaa)
 - XDrawOffset for animations (by Morton)